requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
configparser==3.7.4
entrypoints==0.3
enum34==1.1.6
flake8==3.7.7
funcsigs==1.0.2
functools32==3.2.3.post2
idna==2.8
mccabe==0.6.1
more-itertools==5.0.0
pathlib2==2.3.3
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytest==4.3.1
pytest-sqlalchemy==0.2.1
python-dotenv==0.10.1
requests==2.21.0
scandir==1.10.0
six==1.12.0
slackclient==1.3.1
SQLAlchemy==1.3.1
SQLAlchemy-Utils==0.33.11
typing==3.6.6
urllib3==1.24.1
websocket-client==0.54.0
